--- src/components/Blockquote/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Blockquote';
