+++ src/components/HorizontalRule/index.tsx
