--- src/components/Logo/index.ts
@@ -1,1 +1,0 @@
-export { default } from '@components/Logo/Logo';
