--- src/components/SEO/index.ts
@@ -1,1 +1,0 @@
-export { default } from '@components/SEO/SEO';
