--- src/components/Section/index.ts
@@ -1,1 +1,0 @@
-export { default } from '@components/Section/Section';
