--- src/components/SocialLinks/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SocialLinks';
