--- src/components/Subscription/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Subscription';
